+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Security in case of vulnerabilities.
 ### Changed  
 
 ### Fixed  
+- Optimised queries behind the "By local authority" page
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-05-30.320...main) for everything awaiting release